fix(src): complete LLP64 sweep for calamp/selsbc/selpsc
On Win64 LLP64, `long` is 32-bit while on Linux/macOS LP64 it is 64-bit,
so code that relies on `long` having a specific width breaks under MSVC.
Remaining offenders missed by Task 1.2:

- calamp.cpp: `long unsigned int nof_pixels` / `nof_zero_pixels` are
  32-bit pixel counters; switch to `uint32_t` (add `<cstdint>`).
- selsbc_patch.cpp, selpsc_patch.cpp: the Sun raster magic check used
  `long magic=0x59a66a95` and `*reinterpret_cast<long*>(header)`, which
  reads 8 bytes on LP64 vs 4 bytes on LLP64 -- the header is a 4-byte
  big-endian word, so the cast must be `int32_t*` and the constant
  `int32_t`. `<cstdint>` is already included in both files.

Builds cleanly with cmake; test_magic_fix, test_cstdint_audit,
test_selsbc_return, and test_calamp_format all pass.

+++ src/selpsc_patch.cpp
@@ -266,9 +266,9 @@ try {
     }
 
     char header[32];
-    long magic=0x59a66a95;
+    int32_t magic=0x59a66a95;
     ampfile[i].read(header,32);
-    if (*reinterpret_cast<long*>(header) == magic)
+    if (*reinterpret_cast<int32_t*>(header) == magic)
         cout << "sun raster file - skipping header\n";
     else ampfile[i].seekg(ios::beg);
   }
